ACEfriction.FrictionFit.flux_assemble — Function

flux_assemble(fdata::Array{DATA}, fm::FrictionModel, ffm::FluxFrictionModel; weights = Dict("observations" => ones(length(fdata)), "diag" => 2.0, "sub_diag" => 1.0, "off_diag"=>1.0)) where {DATA<:FrictionData}

Converts FrictionData into a format that can be used for training with ffm::FluxFrictionModel

ACEfriction.FrictionFit.FluxFrictionModel — Type

FluxFrictionModel(c::NamedTuple{model_ids}) where {model_ids}

Create a FluxFrictionModel with the parameters c and the model ids model_ids.

Arguments

c::NamedTuple{model_ids} : The parameters of the model.

ACEfriction.FrictionModels.get_ids — Function

get_ids(m::FluxFrictionModel)

Return the model ids of the model m.

ACEfriction.FrictionFit.l2_loss — Function

l2_loss(fm, data)

Compute the L2 loss of the model fm on the data data.

ACEfriction.FrictionFit.weighted_l2_loss — Function

weighted_l2_loss(fm, data)

Compute the weighted L2 loss of the model fm on the data data.

ACEfriction.FrictionFit.weighted_l1_loss — Function

weighted_l1_loss(fm, data)

Compute the weighted L1 loss of the model fm on the data data.