ACEfriction.MatrixModels.RWCMatrixModel — Type

function RWCMatrixModel(property, species_friction, species_env; 
    maxorder=2, 
    maxdeg=5,  
    rcut = 5.0, 
    n_rep = 1, 
    species_substrat=[]
)

Creates a matrix model with row-wise coupling. By default, this model evaluates blocks $\Sigma_{ij}$ as a function of a spherical pair environment centered at the atom i.

Arguments:

property – the equivariance symmetry wrt SO(3) of matrix blocks. Can be of type Invariant, EuclideanVector, or EuclideanMatrix.

species_friction – a list of chemical element types. Atoms of these lement types "feel" friction, i.e., only for atoms of these element types the matrix model is evaluated, i.e., matrix blocks $\Sigma_{ij}$ are evaluated only if the element types of atoms i and j are contained in species_friction.

species_env – a list of all chemical element types that affect the evaluation of the friction tensor, i.e., blocks $\Sigma_{ij}$ of friction-feeling atoms i,j are functions of exactly the atoms within the pair environemnt (i,j) whose element type is listed in species_env.

Optional arguments:

maxorder – the maximum correlaton order of the ACE-basis. A correlation order of $n$ is equivalent to $n+1$-body interactions.

maxdeg – the maximum degree of the polynomial basis functions.

rcut – cutoff radius of the spherical pair environment.

n_rep – the number of matrix blocks evaluated per atom pair.

species_substrat – a list of chemical element types. At least one atom of such element types must be within the pair-environemt of two friction-feeling atoms i,j in order for the matrix-block $\Sigma_{ij}$ to be non-zero.

ACEfriction.MatrixModels.PWCMatrixModel — Type

PWCMatrixModel(property, species_friction, species_env;
maxorder=2, 
maxdeg=5, 
rcut= 5.0,
n_rep = 1, 
species_substrat=[]
)

Creates a matrix model with pair-wise coupling. In order to allow for good approximation of general friction tensors, this model should be combined with a matrix model of type OnsiteOnlyMatrixModel.

By default, this model evaluates blocks $\Sigma_{ij}$ as a function of a spherical pair environment centered at the atom i.

Arguments:

property – the equivariance symmetry wrt SO(3) of matrix blocks. Can be of type Invariant, EuclideanVector, or EuclideanMatrix.

species_friction – a list of chemical element types. Atoms of these lement types "feel" friction, i.e., only for atoms of these element types the matrix model is evaluated, i.e., matrix blocks $\Sigma_{ij}$ are evaluated only if the element types of atoms i and j are contained in species_friction.

species_env – a list of all chemical element types that affect the evaluation of the friction tensor, i.e., blocks $\Sigma_{ij}$ of friction-feeling atoms i,j are functions of exactly the atoms within the pair environemnt (i,j) whose element type is listed in species_env.

Optional arguments:

maxorder – the maximum correlaton order of the ACE-basis. A correlation order of $n$ is equivalent to $n+1$-body interactions.

maxdeg – the maximum degree of the polynomial basis functions.

rcutbond, rcutenv, zcutenv – Parameters of the ellipsoid-shaped pair environments. rcutbond is the cutoff distance for the distance between the two pairs, zcutenv is the length of the axis (typically this will be the major axis) of the elipsoid aligned with the displacement of atoms i,j, and rcutenv is the length of the axis perpendicular to the displacement of atoms i,j.

n_rep – the number of matrix blocks evaluated per atom pair.

species_substrat – a list of chemical element types. At least one atom of such element types must be within the pair-environemt of two friction-feeling atoms i,j in order for the matrix-block $\Sigma_{ij}$ to be non-zero.

ACEfriction.MatrixModels.OnsiteOnlyMatrixModel — Type

function OnsiteOnlyMatrixModel(property, species_friction, species_env; 
    maxorder=2, 
    maxdeg=5,  
    rcut = 5.0, 
    n_rep = 1, 
    species_substrat=[]
)

Creates a matrix model that evaluates to a block-diagonal matrix. The resulting friction tensor is of the form of a block-diagonal matrix with 3x3 matrix blocks.

Arguments:

property – the equivariance symmetry wrt SO(3) of matrix blocks. Can be of type Invariant, EuclideanVector, or EuclideanMatrix.

species_friction – a list of chemical element types. Atoms of these lement types "feel" friction, i.e., only for atoms of these element types the matrix model is evaluated, i.e., matrix blocks $\Sigma_{ij}$ are evaluated only if the element types of atoms i and j are contained in species_friction.

species_env – a list of all chemical element types that affect the evaluation of the friction tensor, i.e., blocks $\Sigma_{ij}$ of friction-feeling atoms i,j are functions of exactly the atoms within the pair environemnt (i,j) whose element type is listed in species_env.

Optional arguments:

maxorder – the maximum correlaton order of the ACE-basis. A correlation order of $n$ is equivalent to $n+1$-body interactions.

maxdeg – the maximum degree of the polynomial basis functions.

rcut – For row-wise coupled matrix models, the pair environment of the atom pair i,j is by default defined as the set of atoms within a spherical cutoff of radius rcut around the atom i.

n_rep – the number of matrix blocks evaluated per atom pair.

species_substrat – a list of chemical element types. At least one atom of such element types must be within the pair-environemt of two friction-feeling atoms i,j in order for the matrix-block $\Sigma_{ij}$ to be non-zero.

ACEfriction.mbdpd_matrixmodel — Function

mbdpd_matrixmodel(property, species_friction, species_env;
maxorder=2, 
maxdeg=5,    
rcutbond = 5.0, 
rcutenv = 3.0,
zcutenv = 6.0,
n_rep = 3, 
species_substrat=[], 
)

Create a matrix model for a momentum-preserving friction model for the simulation of Dissipative Particle Dynamics. The model is a particular parametrization of a pair-wise coupled matrix model.

This model evaluates blocks $\Sigma_{ij}$ as a function of ellipoid-shaped pair environments centered at the midpoints of the positions of atoms i.j.

Arguments:

property – the equivariance symmetry wrt SO(3) of matrix blocks. Can be of type Invariant, EuclideanVector, or EuclideanMatrix.

species_friction – a list of chemical element types. Atoms of these element types "feel" friction, i.e., only for atoms of these element types the matrix model is evaluated, i.e., matrix blocks $\Sigma_{ij}$ are evaluated only if the element types of atoms i and j are contained in species_friction.

species_env – a list of all chemical element types that affect the evaluation of the friction tensor, i.e., blocks $\Sigma_{ij}$ of friction-feeling atoms i,j are functions of exactly the atoms within the pair environemnt (i,j) whose element type is listed in species_env.

Optional arguments:

maxorder – the maximum correlaton order of the ACE-basis. A correlation order of $n$ is equivalent to $n+1$-body interactions.

maxdeg – the maximum degree of the polynomial basis functions.

rcutbond, rcutenv, zcutenv – Parameters of the ellipsoid-shaped pair environments. zcutenv is half of the length of the axis of the elipsoid aligned with the displacement of atoms i,j, and rcutenv is half of the length of the axis perpendicular to the displacement of atoms i,j. rcutbond is the cutoff for the displacement of the atoms i,j, i.e., if the distance between atoms i,j is larger thant rcutbond, then $\Sigma_{ij}$ evaluates to zero.

n_rep – the number of matrix blocks evaluated per atom pair.

species_substrat – a list of chemical element types. At least one atom of such element types must be within the pair-environemt of two friction-feeling atoms i,j in order for the matrix-block $\Sigma_{ij}$ to be non-zero.